Do not try to append to a single-value field.
This change is opinonated - either it is a list, or not. Do not try to
be more clever than Python itself, I think.

Maybe one could re-introduce this magic by really making LDAPAttribute
its own type acting in certain ways under different circumstances (like
planned, but not achieved before), but there is much more to that (what
if __add__ing a string? or a list?…)